--- pallets/task-scheduler/src/tasks/error.rs
@@ -1,7 +1,0 @@
-#[derive(Debug)]
-pub enum TaskError<E, S> {
-	Evaluation(E),
-	Scheduler(S),
-	/// the task id has been found in onchain storage, this is most likely a dupe.
-	FinishedTask,
-}
