--- pallets/task-scheduler/src/weights.rs
@@ -1,15 +1,0 @@
-#![cfg_attr(rustfmt, rustfmt_skip)]
-#![allow(unused_parens)]
-#![allow(unused_imports)]
-
-use frame_support::{traits::Get, weights::Weight, dispatch::marker::PhantomData};
-
-pub struct WeightInfo<T>(PhantomData<T>);
-impl<T: frame_system::Config> super::WeightInfo for WeightInfo<T> {
-	fn on_initialize(a: u32) -> Weight {
-		(0 as Weight)
-			.saturating_add((10_000 as Weight).saturating_mul(a as Weight))
-			.saturating_add(T::DbWeight::get().reads(1 as Weight))
-			.saturating_add(T::DbWeight::get().writes((1 as Weight).saturating_mul(a as Weight)))
-	}
-}
